Log if we cannot deserialize a line into an airbyte message. (#5948)
After #5136, we removed this block of code as this does not conform to how the Airbyte Protocol is meant to work.

However, since existing connectors do violate this, this PR puts this back for the meantime as a bandaid to give us time to fix this.

Previously, if a line is not a JSON string, the platform assumes this is a log and logs this - we add this back.

Due to the new set up, where for performance reasons we deserialise directly to an AirbyteMessage, what we lose here is the ability to differentiate between a string that isn't a json and a json string that cannot be deserialized into an AirbyteMessage.

Overall, this should result in slightly noiser logs. We would be printing these lines anyway as deserialization errors, so this does not affect performance.