Introduce test connection builder for shorter and safer acceptance tests (#8079)
We have a ton of copy pasta for creating connections in our acceptance tests. This leads to a lot of duplicate code. It also leads to mistakes. Finally, it means we often are "over configuring" connections. For example, we currently run normalization on almost every acceptance test, even though it is relevant in only a small number of them.

In addition, because we were letting the tester set the name and namespace, it means that tests are often writing data into the same tables in a destination which can lead to hard-to-predict behavior.

Also adjusted tests to remove all normalization where it was not needed.

Removed one flaky test. And fixed one test with a race condition.

Co-authored-by: Davin Chia <davinchia@gmail.com>