Simplify Test Harness: Move Asserts out. (#8170)
The Test Harness is huge and complicated. One thing it contains it should not is assertions. Let's move it out!

What I really want to get working is #8050. However that would require some modification to the test harness. After playing around, I decided it would be simpler to reduce some of the tech debt before making modifications.

- Move Assertions out of the Test Harness to an Asserts class.
  - Added Javadocs where I felt was helpful.
- Several renames for more conciseness:
  - Rename AirbyteAcceptanceTestHarness -> AcceptanceTestHarness
  - Rename DatabaseConnectionHelper -> Databases