fix migration: check if json col has object
## What

We added a migration in https://github.com/airbytehq/airbyte-platform-internal/pull/9027 to backfill the `notification_settings` column and enable breaking change notifications.

However, some internal users have run into issues with existing db state. Specifically, with the existence of a literal `"null"` string in the column. This means that the migration fails due to being unable to set a value within the json object.

## How

Fix the migration by checking if the json type is actually an object, so that we're able to set properties on it. If it's not an object, we'll leave it alone.

It also adds tests for the possible values we can have (the literal `"null"` string and just in case, because the col is nullable, actual `null` values)

Note that we're modifying an existing migration here. It's not ideal, but creating a new migration won't solve since the previous migration fails. Cloud has already run this migration (successfully) so it will have no effect there.

**Note**: The existence of these null strings smells like a bug and something that should not be happening.