Fix Custom Transformation Error. (#9367)
Fixes airbytehq/airbyte#31353.

Normalization V2 removed one piece of metadata the custom transformation runner needed to run. This caused an OSS bug users are now reporting.

The various remediation paths are discussed here: https://airbytehq-team.slack.com/archives/C03C4AVJWG4/p1697583573340729

The one we finally picked was:

Continue using the current normalization image to set things up. Instead clarify we are using this to set up DBT. Do so by renaming the image to CustomTransformationPrep.
Extract the correct integrationType from the destination image name. This is required as the custom transformation prep needs to understand how to look at destination config to connect to the destination to run the transformation in the warehouse.
I also took the chance to remove all the Custom Transformation related code from the Normalization Runner to prepare us to delete the Normalization code once 1S1T is done.

All of this is short term fixes to fix this feature for OSS users. I have a follow up item to think about where this feature sits in our product roadmap.