Remove timestamp fields from the workload api. (#9902)
Timestamp field lead to some deserialisation funkiness with Kotlin.

The openapi generator as configured on this commit doesn't recognise timestamps and gives us bad serialisations.

We'll remove it for now because we don't really need this.

We should add this back because the api should return all fields in the database by convention. In that case, we'll add it back as an integer representing epoch milliseconds. This also matches the rest of our api convention.