Oncall: Do not fail destination in case of a malformated log line (#9934)
Pairing with @bleonard on Friday we both learnt that all Java destinations and sources actually log directly to standard out without use of the Airbyte Log Message. The Airbyte Protocol allows for this, though it's strongly not recommended.

With the current logic, the entire sync will fail if a log line contains the 'record' message. E.g. Snowflake currently logs out stream names, so this will fail a job if any of its streams has the name record in it. This is currently happening in prod, which is bad. It must be affecting a small enough group that it is not being reported.

Ideally, we want to get all Sources/Destinations using Log Messages and amend the protocol to allow strict validation. This will take a while, so in the mean time, we should not error at all or risk blowing up syncs with connectors that log any lines containing 'record'.

Update: The git history here is slightly confusing. The meat of the change is #9951. That diff was created as we changed our strategy and I wanted to highlight the diff.

Co-authored-by: Davin Chia <davinchia@gmail.com>