Helm: Better define when to recreate crucial resources. (#10091)
This closes airbytehq/airbyte#32203.

#9469 turned on the orchestrator by default for OSS Kube deployments. Before this, OSS Kube jobs would fail whenever Airbyte is deployed.

When we turned this on, I did not occur to me to test the upgrade path. What our helm charts do is recreate the airbyte-ab and minio pods each time. This would wipe the state bucket, so jobs would not be able to run after an upgrade.

This PR cleans up the airbtye-db and airbyte-minio behaviour with the side effect of fixing this bug.

- Instead of recreating the airbyte db and the minio pod each time, we only create these critical resource on install. Once airbyte is running, there is no situation where recreating these resources on upgrade is needed. In fact, this is harmful since all jobs running at that time will fail. This also slows down the upgrade since these resources are required before the actual Airbyte application can start up.
- Pin minio to a specific version instead of always pulling the latest version. Although we haven't yet seen minio version bug issues, pinning to a specific version provides more stability.
- Do the same for kubectl.